refactor: flatten interpreter gas module
