+++ crates/evm2/src/interpreter/gas.rs
